requirements.txt
# Voice
openwakeword>=0.6.0
faster-whisper>=1.0.0
pyttsx3>=2.90
sounddevice>=0.4.6
webrtcvad>=2.0.10
numpy>=1.24.0

# AI  (httpx is the only hard requirement — used by both Ollama & OpenRouter)
httpx>=0.25.0
python-dotenv>=1.0.0
# openai>=1.0.0  # optional, only if using OpenAI-compatible endpoints

# Browser
playwright>=1.40.0

# Desktop Control
pyautogui>=0.9.54
pyperclip>=1.8.2

# Screen Reading
mss>=9.0.0
easyocr>=1.7.0

# System Control
pycaw>=20230407
screen-brightness-control>=0.22.0
psutil>=5.9.0
comtypes>=1.2.0

# Memory
# sqlite3 is built-in

# GUI
PyQt6>=6.6.0

# Packaging
pyinstaller>=6.0.0
